+++ devnets/mod.ts
@@ -4,10 +4,10 @@ export * from "./binary.ts"
 export * from "./CapiConfig.ts"
 export * from "./chainSpec.ts"
 export * from "./createTempDir.ts"
+export * from "./dev_users.ts"
 export * from "./DevnetConnection.ts"
 export * from "./devnetsHandler.ts"
 export * from "./devnetsUrl.ts"
-export * from "./devUsers.ts"
 export * from "./proxyWebSocket.ts"
 export * from "./startNetwork.ts"
 export * from "./syncConfig.ts"
